DBA-791
Remove emcli_login.sh usage as it stores the password in the script. Added a more secure method by adding a new common task that can be called instead, although not in the same way as bash script within tasks. It should be called at the start of a block that has tasks calling emcli.