DBA-1026
No need for configure_osbws.yml as we now do the reinstall and configuration of OSBWS as part of the prepare-clone call to reinstall_osbws.yml, which now moves the wallet out of the Oracle home and into $ORACLE_BASE/wallets (where other wallets are already being stored). We will always unzip the osbws_linux64.zip file if it exists as that contains the library files. We don't need to run the installer (osbws_install.jar) if the osbws.ora file already exists.