DBA-1026 refactor OSBWS handling in Release Update (#661)
* DBA-1026

No need for configure_osbws.yml as we now do the reinstall and configuration of OSBWS as part of the prepare-clone call to reinstall_osbws.yml, which now moves the wallet out of the Oracle home and into $ORACLE_BASE/wallets (where other wallets are already being stored). We will always unzip the osbws_linux64.zip file if it exists as that contains the library files. We don't need to run the installer (osbws_install.jar) if the osbws.ora file already exists.

* Update oracle-db-release-update.yml

Added combo patch 37952382 in order to test prepare-clone phase

* Update reinstall_osbws.yml

Fixes

* Update reinstall_osbws.yml

Fix for setting the IAM role value as ansible_ec2_iam_instance_profile_role isn't set when running the GHA

* Update reinstall_osbws.yml

Check whether the new wallet file contains Amazon Root CA certs and replace with backup if it doesn't but leave if it does

* Update reinstall_osbws.yml

Remove comment and fix failed_when checks

* Update reinstall_osbws.yml

Simplify playbook to only copy wallet out of the Oracle Home and update the config files in the new Oracle Home to point to it. If it needs to be reinstalled then the oracle-secure-backup role should be used for that.

* Commit changes made by code formatters

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>